+++ src/main/webapp/swagger/v0/changelog.md
@@ -2,6 +2,15 @@
 
 Source: ${grailsApplication.config.grails.serverURL} - Version: ${de.laser.api.v0.ApiManager.VERSION}
 
+#### 0.125
+
+- added new attribute: `Package.scope`  
+- added new attribute: `Package.file`
+- removed attribute: `Package.fixed`
+- removed attribute: `Package.packageScope`
+- removed attribute: `Package.listVerifiedDate`
+- removed attribute: `Package.listStatus`
+
 #### 0.124
 
 - internal refactoring